requirements.txt
pytubefix
youtube_transcript_api
beautifulsoup4
chromedriver-autoinstaller>=0.4.0
selenium>=4.10.0